requirements.txt
torch>=1.3.0
nltk==3.4.5
numpy>=0.13
scipy>=1.2.0
scikit-learn>=0.17
spacy==2.0.12
matplotlib
tqdm
bokeh
numba
gensim
pandas
glove_python
faiss-cpu
prefetch_generator
tensorboard
pandas
redis
